requirements.txt
Django>=4.2,<5.0

gunicorn

whitenoise

google-cloud-storage

numpy
scipy

soundfile
librosa

python-dateutil

python-dotenv
djangorestframework


django-storages[google]



google-auth
google-auth-httplib2
google-auth-oauthlib
